requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-multipart==0.0.12
PyYAML==6.0.2
Jinja2==3.1.6
httpx==0.28.1
spacy==3.7.5
langgraph==0.6.7
langchain-core==0.3.76
qdrant-client==1.11.3
python-docx==1.1.2
pypdf==5.1.0
pdfplumber==0.11.4
sentence-transformers==3.1.1
onnxruntime==1.20.1
optimum==1.23.3
